pyproject.toml
[project]
name = "mock-api-service-airbyte"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faker>=37.8.0",
    "fastapi>=0.118.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
